Moved implementations.
Moved CircleCollisionShape and BoxCollisionShape to CollisionShape
