+++ TwinStick/TwinStick.vcxproj
@@ -120,10 +120,8 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ActorCollection.h" />
-    <ClInclude Include="BoxCollisionShape.h" />
     <ClInclude Include="CameraSystem.h" />
     <ClInclude Include="CBufferTypes.h" />
-    <ClInclude Include="CircleCollisionShape.h" />
     <ClInclude Include="CollisionComponent.h" />
     <ClInclude Include="CollisionShape.h" />
     <ClInclude Include="CubeMesh.h" />
@@ -137,6 +135,7 @@
     <ClInclude Include="GraphicSystem.h" />
     <ClInclude Include="HealthComponent.h" />
     <ClInclude Include="IComponent.h" />
+    <ClInclude Include="Intersection.h" />
     <ClInclude Include="ISystems.h" />
     <ClInclude Include="MeshComponent.h" />
     <ClInclude Include="MovementComponent.h" />
